backend/requirements.txt
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
filelock==3.18.0
fsspec==2025.7.0
gradio_client==1.11.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
packaging==25.0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
safetensors==0.6.2
sniffio==1.3.1
spotipy==2.25.1
sympy==1.14.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.0
typing_extensions==4.14.1
urllib3==2.5.0
websockets==15.0.1
